drbd: Introduce functions to modify a prepared send buffer
resize_prepared_command() can be used to shrink the prepared send buffer.
If we realize before sending that we allocated too many bytes this function
has to be used to shrink it to the right size.

additional_size_command() is used to express that with the next send_command()
the size field in the header should be set to express that additional data
will be sent after send_command.