drbd: Fix unexpected drbd_sync_handshake()
In drbd-8.x context the condition was absolutely valid, since
when the peer role changed to primary it was implicitly true
that the peer's disk becaume D_UP_TO_DATE in the same
state change.

In drbd-9 the peer can become primary by relying on a disk
of a third node. Therefore an explicit test on the
peer's disk state is mandatory.