drbd: Fix the force primary state transition
The force primary state transition has to be seen by the other
members of the cluster as a single transaction.

So far the code missed that and sent only the transition to primary
within the two-phase commit. That the disk got updated was sent
with a subsequent send state. -- That was of course not the intention.