drbd: Move pp_in_use* to connection (from device)
since max-buffers is now a connection property, we need to
account the number of used pages per connection. Not per
device.

drbd_free_peer_req() only needs the peer_request to free,
since that has a pointer to a peer_device, which in turn
links to a connection.

drbd_free_pages() now takes a connection instead of
a peer device.