drbd: Fix send_uuids110() for peer we are syncing from
When receiving resyncs from two peers, where one is UpToDate and one is
Outdated, and the resync from the UpToDate node finishes first, we sent out
a P_UUIDS110 packet that was misleading.

  On the receiving sind of a P_UUIDS110: When one of our peers has a
  bitmap-uuid of 0 towards a third node, and we are in sync with the peers
  we assume that we must be in sync to that third node as well. Therefore
  throwing away any bitmap we might have towards the third node.

The bug was that when we finish the resync from the UpToDate node we
send out P_UUIDS110. For the peer we are still sync target we sent
a bitmap uuid of 0.

Fixed that.