drbd: Queue state transition that have precedence
If a request for a state transition comes in from a node with lower node_id
than what we currently have prepared we know that the currently prepared one
will be aborted very soon.

I do not want to throw it away right away. I want to rely on the abort
message form the initiator of the request.

Up to now we rejected the new incomming request. Queue it instead, since
the one that is in the way will get aborted soon.