drbd: Handle very short resyncs
We need to find an other solution than blocking the receiver in
drbd_start_resync(). This is reverted by the previous commit.

  The reason for that is that the two-phase commit mechanism
  needs the receiver to recieve either the P_TWOPC_ABORT or the
  P_TWOPC_COMMIT packets. It it is blocked in the
  down(&device->resource->state_sem) in drbd_start_resync()
  everything gets locked up.

Apparently the problem this commit tired to solve is that
a state packet could signal that the resync is finished comes
in while we still postpone starting the resync.

This patch implements an alternative that tries to find all
that out while receiving those P_STATE packets.