drbd: Fix twopc_phase2() messing up TWOPC_PREPARED bits
Before this patch the twopc_phase2() function uses and re-sets to zero the
TWOPC_PREPARED connection flag. One is allowed to use these bits as long as
one owns resource->remote_state_change.  This was not the case for last
callsite of twopc_phase2() in change_cluster_wide_state().

The effects of this bug was that a following two-phase commit that uses
__cluster_wide_request() to send it's TWOPC_PREPARE or TWOPC_ABORT
packets (via the old twopc_phase2() function), sends the packet to
less receipients than intended.

Which in turn leads to hard to understand effects of two-phase commits that
only get terminated by the timeout.

Fix that by no longer messing with the TWOPC_PREPARE bit in twopc_phase2()