drbd: External retry loop for C_CONNECTING -> C_CONNECTED
The state transition from C_CONNECTING -> C_CONNECTED is done by the
sender. While a sender item is executed the transport is corked.

In case the transaction gets aborted, the internal retry of
change_cluster_wide_state() sends the P_TWOPC_ABORT packet, sleeps and tries
again. But the packet gets never sent since the transport is corked.

Improve that by switching that to an external retry loop, that
returns control to the send_worker while waiting. This in turn
causes the transport to get uncorked and the P_TWOPC_ABORT packet
gets sent in a timely manner.