drbd: drbd_req_destroy expects resource->req_lock to be held
We used to call drbd_req_destroy from do_retry() without the req_lock held,
because the only action that would have required that lock was list_del()
of the request from the transfer log, and requests processed in do_retry
are already on a local, only caller visible, on-stack list.

But meanwhile, drbd_req_destroy does several other actions
that expect the req_lock to be held. So grab it from do_retry() as well.