drbd: Fix too early end of resync on SyncSource side
From drbd-8.4 came the idea that we can detect the end of a resync
from the P_STATE packets. In DRBD-9.0 there are many other reasons
we we might get a P_STATE packet.

The issue was that we might get in a P_STATE packet due to some
unrelated reason, that causes a call to drbd_resync_finished() too
early.

Circumvent that be making sure that we actually saw that the
other node started the resync.