drbd: Fixed two-phase-commits without abort-timer
So far the code only started the timer if the requested state change
could be prepared successfully locally.

That is of course wrong, we need to start it no matter if the local
prepare was successful or not.

A coordinator might die independently from the fact if I was locally
successful in preparing the transaction.