drbd: Fix resync decission after primary crash
This is actually a subtle regression introduced with generating a
new current UUID with late (i.e. with the first write).

Due to that change a new primary does _not_ generate the new UUID
immediately. Now consider a former primary, that was crashed
rejoins the newly promoted primary. On the newly promoted primary
not a single write was done so far.

==> These two nodes have the same current UUID !!

The code as it was assumed that this is a common power failure
and used the CRASHED_PRIMARY flags to determine a resync direction.
The result of that is that it wants to resync from the crashed
primary to the current primary. That is prevented later on.

The solution is to realize that this is not a common power
failure, but to base the decision about the resync direction on
the current roles.

The patch is quite big, since it was necessary to insert
new values into the return values of drbd_uuid_compare()