drbd: Keep twopc_timer armed
The idea to deactivate the twopc timer when getting all expected
packets from nested transactions in was simply wrong.

The timer function will re-shedule itself until it manages to
queue the abort work_callback.

Testing if the list of the work_item is empty is not sufficient.
We might have a race where the worker removes the item from the
list, and then calls the callback.

  With that we had possible race conditions:
    An other context sees it already removed from the list
    and sets the callback to something else. The original
    callback executes in parallel and sets the .cb to NULL
    at its end.
    The other conetext queues the item again. -> When the
    worker executes it we crash with RIP = NULL.

Better test if the .cb is NULL and set the .cb while holding
the req_lock.