drbd: Do not do state transition repl( WFBitMapS -> Established )
In case the peer tells us the it is L_ESTABLISHED while we think that we are
in a resync operation or starting a resync operation, do not update our
repl state.

This minor bug can eventually lead to a BUG() statement when
adding the peer_device->start_resync_timer twice.