drbd: fix for 8.4 <-> 9.0, writes on 8.4 during resync deadlock
In drbd 8, we do activity log transactions in normal operation
on the sending (Primary) side only, and on the receiving side
only in case the sending side has no disk
(e.g. detached after iO error).

In drbd 9, the receiving side always uses the activity log as well,
which can cause conflicts with a simultaneously on-going re-sync.

In DRBD 9, we can "cancel" resync requests, but the compat protocol
to DRBD 8 does not allow that.

To avoid the resulting deadlock potential, we will not use the activity
log on the receiving side in DRBD 9, if the peer is DRBD 8, unless the
peer is diskless.