drbd: fix bitmap op buffer overruns (random memory corruption)
With EXTRACT or MERGE, which are used during bitmap exchange
with "use-rle off", or if that chunk of bitmap happens to be
"uncompressible", our bm_op() did not properly check the
end-of-buffer condition, and would overrun the input/output
buffers by 4 byte, corrupting "random" pages.

Depending on where those 4 byte are, that may even go unnoticed
for some time, or seriously confuse the system, corrupt data,
or crash applications.

The symptom we have seen most frequently:
the oom-killer was triggered, and basically killed off everything.