drbd: don't ignore peer device size during handshake
During handshake, the connection is not yet established.
If we filter out peer device sizes from not-established connections
that defeats the purpose of the handshake.

The result was that DRBD peers "agreed" to their respective own device
size, regardless of the peers device size, where they should agree
on the minimum device size of all peers.

Fix: drop that filter.