Makefile: provide a km-deb target
This eases Debian package building for users that try to build a Debian
package (m-i or dkms) from a fresh git checkout.

Currently, building the dkms package succeeds, but at install/build time
it fails because drbd/drbd_buildtag.c does not exist (it is auto
generated) and at dkms build time it is not possible to generate it
because we are no longer in a directory that contains a .git directory.

The new target makes sure that drbd/.drbd_git_revision is generated.
This is enough for the later generation of drbd/drbd_buildtag.c

In case a tarball containing a debian directory is used, the
regeneration of drbd/.drbd_git_revision itself fails (because there is
no .git directory), but the test for drbd/.drbd_git_revision succeeds
because the file is included in the tarball. Therefore, the new target
can be used in release tarballs (containing a debian directory) and for
fresh git checkouts.