Workaround flow-control deadlock
Right now flow-control messages are only received when drbd's receiving thread
is actively blocked in one of the transport's receive functions.

Apparently in order to not block the number of rx_descriptors (and
tx_descriptors) need to be high enough to have room for the largest possible
BIO.

A better approach would be to receive already in the callback
(dtr_rx_cq_event_handler()). Then we could process the flow-control
messages right away. Then probably it would also work with
smaller numbers of tx_descs and rx_descs.