drbd: fix access after invalidate of packet_info
This fixes two bugs.

(1) In receive_Data, we have a packet_info pointer,
pi->data points into some receive buffer owned by the transport.

As soon as we call into the transport again to receive further data,
e.g. to receive the bio pages with read_in_block(), pi->data has to be
considered invalid.

Most of the time it will still contain the same data, but sometimes it
will have been already freed, recycled, and/or re-used.

The impact was that, sometimes, the not-yet copied to on-stack values
(dp_flags) would appear to be changed, and the peer_request would be
submitted with the wrong flags, or "forget" to send a P_WRITE_ACK on
completion, resulting in potential data corruption, an apparently hung
DRBD protocol, or "BARRIER_ACKED but pending" asserts.

(2) If receive_Data does no longer have a local disk,
for P_TRIM it would have send a P_NEG_ACK with wrong size.

This was noticed while fixing (1), and is fixed as a side effect.

Fix is to stow away the necessary details in a new
struct drbd_peer_request_details before calling read_in_block.