compat: fix in_flight io accounting again
In older kernels, in_flight is a plain int.
We use start IO accounting outside of locks,
which means if adjustments to in_flight happen quasi-simultaneously on
from more than one context, we may be bitten by the read-modify-update
cycle of plain int ++, in_flight will become non-zero even during idle
periods, and the whole iostat accounting will be wrong.

We used to test for this, and explicitly add a spinlock around the
start IO accounting. (Our end-io accounting happens inside the spinlock
anyways). Recent patch to use (and backport) generic_start_io_acct()/*_end_*
dropped this compat-spinlock, introducing a regression for iostat
accounting for kernels older than 2.6.39.

atomic_t is struct { int counter }, this is *our* gendisk, it is not
partitionable, we are the only ones allowed to update this counter.
I chose to not re-introduce the spinlock, but cast the "int in_flight"
to "atomic_t" instead.

The other stats are per-cpu variables anyways.

Which may still leave room for mis-accounting, if an end-io accounting
from IRQ context races with a start IO accounting from process context?
That would be a problem of linux block layer core, though.