drbd: Complete connect logic for primary & diskless nodes
When a node with a current UUID different to our exposed data UUID tries to
connect to a primary & diskless node the latter denied to attempt up to now.
In a pre drbd9 world this was the correct behavior.

Since drbd9 can handle multiple connections, it makes no longer any sense
to reject such an connection attempt.