drbd-km.spec: Port to openSuSE
Currently building an RPM with the drbd-km.spec fails on openSuSE
because of an incorrect build dependency (kernel-syms instead of
kernel-default-devel). Fix the algorithm in the spec file for
determining the kernel development RPM name. Note: if the kernel
development headers have not been installed as an RPM the
%{kernel_devel_rpm} variable will be empty.

This patch has been tested on openSuSE 13.2, CentOS 7.1 and SLES11 SP3.

Signed-off-by: Bart Van Assche <bart.vanassche@sandisk.com>
Signed-off-by: Lars Ellenberg <lars.ellenberg@linbit.com>