build: fix make -j in initial git clone
Make sure the git submodules are available
before trying to build the kernel module.

When building module and "tools" in the same invocation,
build tools last, so the hint to separate drbd utils
won't get lost in the noise from the module build.

Don't try to re-build Makefiles, they are not generated.