fixup: drbd: allow parallel flushes for multi-volume resources
During merge of various attempts of the above patch, I managed
to wait for an uninitialized, otherwise unused, on-stack completion
in drbd_flush_after_epoch(), instead of the intended &ctx.done.

This apparently went unnoticed because most of the time, the timing
would cause wait_on_completion() to be skipped.  Also most setups
disable these flushes anyways, and the whole function would be skipped.
Or, the uninitialized on-stack value would happen to contain ->done == 1
already.  This all could lead to deadlocks, or crashes due to access of
uninizialized pointers.