build: fix check_changelogs_up2date
We used to go "$VERSION" -> "$VERSION+1"rc1 "$VERSION+1" (final),
but typically 1.2.3rc1 sorts higher than 1.2.3 (final).

Embed the "rc" and similar tags not in the "version" but in the
"release" part of *.spec and debian/changelog, and fix our sanity check
make file target to understand that.

Maybe we should instead generate all of these
from a "bump release" script, but that is left for an other day.