drbd: Only allow deleting of the last path if we are C_STANDALONE
Deleting the last path while we are connecting is a awkward corner
case for all transports. (Since it leads to the implicit consequence
that the cstate would need to change from C_CONNECTING to C_STANDALONE)
Better do not allow it on a generic level. The user should rather use
disconnect before removing all paths.

With that change also introduce that a disconnect call waits
until the connection reached C_STANDALONE state.