fixed resize for multiple nodes
- Allow resize in clusters where all nodes are secondary
- Fix code for clusters with more than two nodes.

This bumps the protocol version. If the agreed version is too old, a
primary is needed and the old way of using the RESOLVE_CONFLICTS flag is
used. If the agreed version is high enough, make the primary (if it
exists) the sync source. If there is no primary for a connection (or no
primary at all), make the node with the lower ID the sync source.

Usually, nodes get daisy chained by ID while syncing:
A(0,S) -> B(1,S) -> C(2,S) # Node(ID, role)

In case a primary exists, this might change, e.g:
A(0,S) <- B(1,P) -> C(2,S) # Node(ID, role)

In case of a node D(3,S), it would be chained to C(2,S)