drbd: drbd_req_destroy: don't kref_sub instances from an unrelated object
We want to destroy (write) requests in the order we created them
on our resource transfer log.

For each new requests, we get an extra kref
if there is still a preceding WRITE request on the list.

During destruction of a write request, if we find a later write request
on the list, we know "implicitly" that because of the above we have to
drop that extra reference on that request, too.

To avoid recursion in this destructor function,
this is done via explicit tail recursion.

This can easily cause chained destruction of a number of requests.
The assumption is that many of those are for the same "device" aka
"volume".

Previously, the code seems to have assumed that *all* such requests
would be for the same volume, but that is wrong: we have all requests
to all volumes of one "replication group" (resource) on this single
per resource transfer log list.

For a multi-volume resource, this will eventually erroneously kref_sub
references from an unrelated volume, with all strange consequences
and resulting crashes.

This patch lets the tail recursion accumulate several device refs
to be put, but will process them and restart accumulation if the
next request to be destroyed was for a different device.