drbd: Fix transferlog list walks for drbd_req_destroy() with recursion
A list_for_each_safe() works if the current element gets removed
from a list while walking it. The issue was that a _req_mod()
might destroy a request. That in turn might destroy other requests.
If one of the other requests is the safed element we are doomed.

The solution is to keep a reference on the safed (or next)
element while walking the list.