drbd: debugfs, peer_device: add 8.4 style '/proc/drbd' equivalent
This may still be re-formatted, and is in part still incomplete,
it is also not exactly the same format.

Congestion reason is missing, write ordering is given as numeric,
it shows pe:[AP;RS] and ap:[WRITE;READ] separately,
and always behaves as if "proc_detail" was turned on.

Not sure if "proc_drbd" is the best name,
if we really want to keep it, we probably want to rename it again.
But for now it is good enough.