drbd: limit submission io-depth (nr_requests)
We need at least some way to limit the number of "in-flight" requests
aka the submit io-depth.
Otherwise, given a sufficiently braindead user, we'd happily queue up
incoming requests until memory or other system resources are exhausted.

In drbd.conf, or on drbdsetup command line,
we call this resource option max-io-depth.
In kernel, it is res_opts.nr_requests in analogy to request_queue.nr_requests.