drbd: Create new current UUID if peer's disk fails
So far the code waited until a new write came in after a
peer's disk failed.
That was wrong since the write that triggered the failure
already created the difference. In that case create the
new current UUID as soon as possible.

What stays unchanged:
In case the peer_disk goes to D_UNKNOWN delay generating
the new UUID until a new write comes in or a previous
(temporarily frozen) write gets completed.