drbd: Extend the semantic of auto-promote-timeout
So far it was only used wait that
 * a resource becomes secondary an other node
 * a device in such a resource is closed on an other node
during trying an auto promote action.

With this patch it extended to wait for a D_UP_TO_DATE disk
(locally or remote) in case there is no up-to-date disk
right now.

Before that patch, DRBD failed the open immediately in case
there was not up-to-date disk present.