drbd: Fix a BUG by no longer add_timer() twice
What was happening quite infrequent:

in drbd_connect() we arm the connect_timer. Then assume a down command comes
in and we go into conn_disconnect(). Which deletes the timer and flushes
the sender workqueue.

Let's assume the timer was able to trigger and queued the connect_work()
worker callback function. When that is executed during the flush
sender workqueue it tries to change state to CONNECTED and probably
fails with SS_CONCURRENT_ST_CHG. I.e. it re-arms the timer!

Later on the user whishes to re-establish the connection. Boom the
timer was already added!

The fix is obviously, when the work function finds the cstate in
something else than C_CONNECTING it drops the connection reference
and exits immediately.