drbd: Fix using uninitialized reply->primary_nodes
There is a number of callsites in the code that do a strate transition to
C_DISCONNECTING with the CS_HARD flag:
change_cstate(connection, C_DISCONNECTING, CS_HARD);

in change_cluster_wide_state() such requests get handled as local-only
state change. I.e. the change() function (in that case do_change_cstate() gets
called with prepare = false.

The do_change_cstate() function in turn looks at reply->primary_nodes, where it
might find a mask of primary nodes from the previous two-p-c transaction.
That might lead to unexpected outdate of the local disk.

Fix that by:
Have a enum change_pase as argument to dhe change() function. With that
each of those callbacks can find out if it is called for a local
transaction or if it is the commit of a cluster wide transaction.
I.e. do_chnage_cstate() does only do the local outdate action if it
knows that the reply->primary_nodes is valid.

The issue was introduced with:
32a8b32 drbd: Outdate a node when it is gracefully disconnected from a primary...