drbd: Fix how a D_OUTDATED node gets D_UP_TO_DATE after loosing the primary
In a setup with 3 nodes forming a chain: A - B - C
consider A is R_PRIMARY and D_DISKLESS -> C is D_OUTDATED.
When A goes away ungracefully B becomes temporarily D_CONSISTENT and
goes back to D_UP_TO_DATE as soon as it knows that there is not
different path to the primary.
C does a resync-after-unstable with B. If there is no resync it
need to take over the D_CONSISTENT disk state, which will go over
to D_UP_TO_DATE soon.