drbd: fix for possible deadlock in kernel >= 4.3
To avoid very long internal queues as we wait for our replication
peer device and local backend, we limit the number of in-flight
bios we accept, and block in our ->make_request_fn() if that
number exceeds a configured watermark ("max-buffers").

linux mainline commit
54efd50 block: make generic_make_request handle arbitrarily sized bios
introduced blk_queue_split(), which will split any bio that is
violating the queue limits into a smaller piece to be processed
right away, and queue the "rest" on current->bio_list to be
processed by the iteration in generic_make_request() once the
current ->make_request_fn() returns.

Before that, any user was supposed to go through bio_add_page(),
which would call our merge bvec function, and that should already
be sufficient to not violate queue limits.

Previously, typically the next in line bio to be processed would
be the cloned one we passed down to our backend device, which in
case of further stacking devices (e.g. logical volume, raid1 or
similar) may again push further bios on that list.
All of which would simply be processed in turn.

Now, with blk_queue_split(), the next-in-line bio would be the
next piece of the "too big" original bio, so we end up calling
several times into our ->make_request_fn() without even
dispatching one bio to the backend.

With highly concurrent IO and/or very small "max-buffers" setting,
this can deadlock, where the current submit path would wait for
the completion of the bios supposedly already passed to the
backend, but which actually are not even dispatched yet, rotting
away on current->bio_list.

I see at least five possible fixes:

a) involve yet an other kernel thread or work_queue, and punt
all bios we find on current->bio_list that are not to ourselves
over there before we block

b) don't block, if we have already at least on in-flight bio
and current->bio_list is not empty

   (one possible)
   implementation: add to the top of drbd_int.h  may_inc_ap_bio():

	/* To make progress, we may need to have current->bio_list
	 * be drained by the iteration in generic_make_request().
	 * If we have already some bio in-flight, and ->bio_list
	 * is not empty, ignore any condition below, just proceed. */
	if (current->bio_list
	&& !bio_list_empty(current->bio_list)
	&& atomic_read(&device->ap_bio_cnt) > 0)
		return true;

c) if on entry we find bios not to ourselves on current->bio_list,
reorder that list so all bios not to our device come first, then
the current incoming bio, then any alreay queued bios to ourselves.

I dislike a), and b) may overshoot the configured "max-buffers"
setting by several thousand, depending on IO concurrency,
read-ahead settings and other factors; it also may introduce
problems when ignoring the suspended, stable-state, or other non
in-flight watermark conditions we may block on.

With c) we have a problem if some other driver implements similar
hackery.

Of course there is d), fix the generic block layer, which is currently
being discussed here: https://lkml.org/lkml/2016/6/22/172

Then there is e), which is implemented here:
temporarily reset current->bio_list to NULL to fake initial
entry and thus have one level of real recursion.