drbd: Fix abort of two-phase-commits
In a three node chain A - B - C lets consider a A is the coordinator or
a transaction (it begins there), and C sends a P_TWOPC_NO back.
With the code as it was, B aborted the transaction when it forwarded the P_TWOPC_NO
back to A.
The problem is that when the coordinator sends the P_TWOPC_ABORT, it is ignored
on B (since it has no knowledge of the transaction any more). B does not forward
the P_TWOPC_ABORT to C. C has the transaction until it times out :(

Fix it by keeping the transaction on B in the moment we send the P_TWOPC_NO back.

One might be tempted to consider to throw away the connection on C in the
moment sending back the P_TWOPC_NO. But then consider this for a moment:

       C
      /
 A - B
      \
       D - E

A .. coordinator
D .. can not do the transaction

The transaction never reaches E, but it reaches C. Forgetting it early
on D and B is okay, but C would not see the P_TWOPC_ABORT packet.

--> Better keep it everywhere until P_TWOPC_ABORT is sent out by the
coordinator.