drbd: Check socket state instead of checking for readable data
there might be bytes to read on a socket that is already disconnected from
the other side. Better check the socket state.

On the DATA_STREAM this made us blind to already disconnected sockets,
since the feature packets was already in the socket, but there is a short
timeout behind the feature packet...