drbd: Disconnect sockets before releasing them
In numerous places the code simply released a socket when
we realized that something went wrong. That simply destroy
the local socket. It does not care to notify the peer
that a tcp connection might get broken by that.

The peer finds out about that when it sends something
into the connection. It gets than an EPIPE back.

Much more gracefull is to really close the connection before
freeing the socket. With that the peer recognises a lot
earlier that a tcp connection no longer exists.