drbd: two-phase resize
Extend the two-phase commit protocol used for state changes to support resize
operations. This is introduced with protocol version 112.

In phase one (prepare) all the nodes agree to the transaction as long as all
peers support protocol 112. With the reply they send their maximal available
storage space back.
The nodes that initiated a nested transaction send the minimum of the
sub-transaction's participants back.
At the initiator of the transaction the minimum of the whole cluster
arrives, it determines the new size for all nodes and commits the
transaction.