drbd: Do not take any locks in drbd_destroy_connection()
Since the call sites may be anywhere, having arbitrary locks. All the
peer_requests and the socket can already be destroyed when removing
the connection from the data structures.

If in __clear_remote_state_change() the connection is actually freed,
a spin_lock recursion happened with the code as it was. It is very likely
that exactly this was triggered with tests that do disconnect and
two-phase resizes concurrently.