drbd: Fix two-phase-commit when we get the prepare packet multiple times
In a cluster like that:

    1
   / \
  0   3 - 4
   \ /
    2

when a transaction is started by node 0, the prepare packet arrive
twice at node 3.

Send the previous prepared reply, if the reply was ready.

If the reply is not yet ready make sure it gets sent to all parents
when it becomes ready.

Previously we sent an incomplete reply to the node that sent us
the request the second time.
This bug might caused various ill effects for state transitions, that
are not known. It became visible since two-phase-commit are also used
for resizes.