drbd: do the resize after relaying the commit message
A resize operation may take some time. I.e. when we local pending IO needs to
drain, or when allocating the new bitmap causes write-out of dirty pages.
Since we want that the new size becomes available on all nodes as soon as
possible, it is better to to relay the commit messages to all nodes taking
part in the transaction _before_ doing the resize on the node.