drbd: Allow peer_devices without bitmap
It does not make sense to allocate a bitmap for nodes with are
prepmanently diskless.
The peer-device-options command got a new option "bitmap". When it
is set to false, no bitmap index will be allocated for this peer_device.

In order to avoid the bitmap allocation user-space has to create
the minor, then the peers and finally set the peer-device-options
to avoid the bitmap slot allocation.

Either an attach or an connect operation will allocate a bitmap
slot if now avoided by setting the peer-device-option bitmap to false.