drbd: fix outdate_on_disconnect, was too eager to outdate itself
If we "cleanly" disconnect a replication link in a coordinated fashion
via some "cluster wide state transaction", and that link connected some
secondary with some primary, we intend to politely ask the secondary
to outdate itself.

Which means that, on the to-be-disconnected connection,
if our roles differ, and we are in primary role,
ask the peer to outdate itself.
If the peer is primary, and we disconnect,
tell the peer that we intend to outdate ourselves.

Previously this effectively self-outdated when "disconnecting" (stopping to try
to establish a connection) a not fully established connection on a Secondary.