drbd: don't ignore fencing policy just because we have no recorded changes
Just becaue we don't know of any changes wrt. our peer,
that does not mean the peer has not evolved either.

If fencing policies are configured, and we don't know for certain
that we have the "best possible" data, we must stay "Consistent",
and can not implicitly promote ourselves to "UpToDate".