drbd: Enable sending writes to SyncTarget servers
A client (diskless primary) that has two servers (secondary with disk).
If the two servers perform a resync, the client should send its
writes to both servers.

Before the fix the resync target considers itself as D_UP_TO_DATE
after the resync, _BUT_ it missed the writes that happened during
the resync!! This was a possible source of data corruption!!

BTW, on the snyc source server these blocks where correctly marked
as out of sync to the sync target, but if those marks arrive in
the bitmap "after" the current-resync-position then they stay
there in the bitmap.