drbd: Fix race that lead to data-inconsistency in the past
When two nodes start a resync and a client (diskless primary) is connected
to both of them, then the client may send its modifications only to the
node that becomes sync source. The reason is that the future sync
target is D_OUTDATED, and such nodes are not updated by clients.
After the resync started, the client will send its writes to both nodes
(since the sync target is D_INCONSISTENT).

The problem is the writes that happen in the moment the resync gets
started. In order to solve that:

1) The sync source sends P_OUT_OF_SYNC packets to the target when
   it adds bits to the bitmap.
2) Make sure the sync target includes such late-set bits in the
   resync.