drbd: Give up local "wenr" references if necessary
The wenr (wait resync extent number) is a concept from the two node cluster
world. In case we find offending references only by the wenr mechanism,
give them up.

This is necessary to avoid deadlocks in a 3 node configuration.

When write originate from node A and B and C are doing a resync at
the same time. (Resync from B to C). B and C get the write into their
AL. C stops to send resync requests.

B gets a first write from A, which makes it into its AL. Then a resync
request gets in, it can not proceed since the entry in the AL, but the
entry is kept and hold by wenr.

If a second write comes in from A that blocks due to the wenr. Nobody
retry the RS side since C stopped sending requests...

Get rid of the wenr reference eary...