drbd: Fix a bug that can lead to IB_WC_RNR_RETRY_EXC_ERR
When we receive a flow control message, we decrease the
rx_descs_known_to_peer (but not the rs_descs_posted).

The bug was was that we evaluated first if we should send a flow control
message to the peer, and _THEN_ repost the rx_desc of the flow control
message we are currently receiving.

I.e. these events could have happended

A --many-data-packets--> B
A <-flow-controls-pkts-- B
A --flow-ctrl-to-early-> B
A <----something-------- B
A ----RNR NACK---------> B
A finally re-posts; too late; B bails out with IB_WC_RNR_RETRY_EXC_ERR

Fix that by re-posting first, and then consider to send a flow control
message.