drbd: Fix a potential bug that can lead to IB_WC_RNR_RETRY_EXC_ERR
In the dtr_handle_rx_cq_event() function rx_descs_posted and
rx_descs_known_to_peer gets decremented. As each is in itels an
atomic operation the order they appear on other CPUs in not
determined.

dtr_send_flow_control_msg() might execute on an other CPU at
the same time, it reads the two values and acts on the difference.

Order how the two operations become visible, order the two
reads. In case they execute in parallel, posted might be already
decresed and known gets decreased later. We might get a -1
due to the reace. The max() masks that and makes it a zero.

The old code might have lead to a +1, which causes the peer
to send data (before we posted a tx_desc) which might case
a IB_WC_RNR_RETRY_EXC_ERR